Improve dev mode performance
