+++ packages/toolpad-core/package.json
@@ -34,7 +34,6 @@
     "url": "https://github.com/mui/mui-toolpad/issues"
   },
   "dependencies": {
-    "@mui/x-data-grid-pro": "^5.17.6",
     "@tanstack/react-query": "^4.10.3",
     "quickjs-emscripten": "^0.21.0",
     "react-error-boundary": "^3.1.4"
