+++ packages/toolpad-core/src/index.tsx
@@ -1,4 +1,4 @@
-export type { PlaceholderProps, SlotsProps, NodeRuntime } from './runtime';
+export type { PlaceholderProps, SlotsProps, NodeRuntime, NodeErrorProps } from './runtime';
 export { Placeholder, Slots, useNode } from './runtime';
 
 export type FlowDirection = 'row' | 'column' | 'row-reverse' | 'column-reverse';
