+++ packages/toolpad-core/src/runtime.tsx
@@ -1,6 +1,4 @@
 import * as React from 'react';
-import { Error as ErrorIcon } from '@mui/icons-material';
-import { Tooltip } from '@mui/material';
 import { ErrorBoundary } from 'react-error-boundary';
 import { RUNTIME_PROP_NODE_ID, RUNTIME_PROP_SLOTS } from './constants.js';
 import type { SlotType, ComponentConfig, RuntimeEvent, RuntimeError } from './types';
@@ -53,10 +51,8 @@ function PlaceholderWrapper(props: PlaceholderWrapperProps) {
   );
 }
 
-export interface NodeRuntimeWrapperProps {
-  children: React.ReactElement;
-  nodeId: string;
-  componentConfig: ComponentConfig<any>;
+export interface NodeErrorProps {
+  error: RuntimeError;
 }
 
 export interface NodeFiberHostProps {
@@ -64,29 +60,6 @@ export interface NodeFiberHostProps {
   [RUNTIME_PROP_NODE_ID]: string;
   componentConfig: ComponentConfig<any>;
   nodeError?: RuntimeError | null;
-}
-
-interface NodeErrorProps {
-  error: RuntimeError;
-}
-
-function NodeError({ error }: NodeErrorProps) {
-  return (
-    <Tooltip title={error.message}>
-      <span
-        style={{
-          display: 'inline-flex',
-          flexDirection: 'row',
-          alignItems: 'center',
-          padding: 8,
-          background: 'red',
-          color: 'white',
-        }}
-      >
-        <ErrorIcon color="inherit" style={{ marginRight: 8 }} /> Error
-      </span>
-    </Tooltip>
-  );
 }
 
 // We will use [RUNTIME_PROP_NODE_ID] while walking the fibers to detect React Elements that
@@ -99,7 +72,19 @@ function NodeFiberHost({ children }: NodeFiberHostProps) {
   return children;
 }
 
-export function NodeRuntimeWrapper({ nodeId, componentConfig, children }: NodeRuntimeWrapperProps) {
+export interface NodeRuntimeWrapperProps {
+  children: React.ReactElement;
+  nodeId: string;
+  componentConfig: ComponentConfig<any>;
+  NodeError: React.ComponentType<NodeErrorProps>;
+}
+
+export function NodeRuntimeWrapper({
+  nodeId,
+  componentConfig,
+  children,
+  NodeError,
+}: NodeRuntimeWrapperProps) {
   const resetNodeErrorsKey = React.useContext(ResetNodeErrorsKeyContext);
   return (
     <ErrorBoundary
